requirements.txt
PySide6>=6.7.0
opencv-python>=4.9.0
numpy>=1.26.0
Pillow>=10.0.0
pytesseract>=0.3.10
mss>=9.0.1
pywin32>=306; platform_system == "Windows"
requests>=2.32.0
beautifulsoup4>=4.12.0
lxml>=5.2.0
